fix(themes): undo regression in themes
